+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## v0.4.0 - 2021-03-xx
 
 - new command: `profile` for generating taxonomic profile from search result
+- `compute`:
+    - new flag `-B/--seq-name-filter` for filtering out unwanted sequences like plasmid.
+    - new flag `-N/--ref-name-regexp` for extracting reference name from sequence file.
+- `search`:
+    - change default threshold value
 
 ## v0.3.0 - 2021-03-16